Introduce "gpu_mem_headroom" to allow adaptive KV cache size.
When the user specifies "gpu_mem_headroom", we will calculate the number
of KV pages based on the available GPU memory after accounting for the
reserved percentage of GPU memory specified through "gpu_mem_headroom".

If the user also specifies "max_num_kv_pages", we will use the smaller of
the two values.