Simplify backend termination logic.
The shell should wait for both the uv process and the backend process to
exit before shutting down the engine.

In the past, we send the SIGTERM signal to both the uv process and the
backend process but we only wait for the uv process to exit.

Now, we send the SIGTERM signal only to the uv process, and we rely on
the uv process to forward the signal to the backend process. We now still
wait only on the uv process to exit, since the uv process will exit when
the backend process exits.

Ideally, we should just use the Tokio methods to send the SIGTERM signal
to the uv process, but unfortunately, the Tokio library currently only
supports sending the SIGKILL signal, which cannot be forwarded to the
backend process and causes immediate termination of the uv process,
which in turn let the backend processes run forever.