Fix waiting for inferlet to finish in the CLI.
Previously, the CLI would wait for the inferlet to finish by detecting
the change in the number of running and finished inferlets. However,
this approach is fundamentally flawed because it cannot distinguish
different inferlets, thus cannot tell which inferlet has finished when
two or more inferlets are running.

The new approach is to wait on the `Instance` object returned by the
engine when submitting the inferlet.