Use internal channels to request the server to stop backend heartbeats.
The `pie` CLI previously used a client to request the server to stop
backend heartbeats. However, the CLI and server components are both
internal to `pie`, so letting the CLI part create a client which in turn
connects to the server through a public websocket complicates the code.

This commit refactors the synchronization between the CLI and server by
using internal channels between them.