Fix the bug that causes ctrl+c to not work with `pie run`.
The `rustyline` library captures the Ctrl+C signal. It works with
`pie serve` because it runs in an interactive shell, and it prints a
message when Ctrl+C is pressed so that the user will know to press
Ctrl-D to exit.

However, `pie run` does not run in an interactive shell. Previously,
the code reused the `rustyline` printer for inferlet output, but that
prevented the Ctrl+C signal from killing the engine process.

This commit fixes the bug by changing the printer to be optional. When
running `pie run`, the printer is `None`, and the code prints the output
through the standard `println!` macro. Since the `rustyline` printer is
no longer created for `pie run`, the Ctrl+C signal is not captured, and
the engine process can be killed by Ctrl+C.